requirements.txt
asgiref==3.5.2
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-db==0.0.7
django-extensions==3.1.5
django-heroku==0.3.1
django-phonenumber-field==6.1.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
entrypoints==0.3
executing==0.8.3
idna==3.3
ipykernel==6.5.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
oauthlib==3.2.0
parso==0.8.3
phonenumbers==8.12.48
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
pywin32==302
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
tornado==6.1
traitlets==5.2.2.post1
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==6.1.0
